# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Easy Way Agency Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface AdminMe {
  userId: string;
  /** @nullable */
  email: string | null;
  isAdmin: boolean;
}

export interface Partner {
  id: number;
  name: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  websiteUrl: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  industry: string | null;
  featured: boolean;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePartnerInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  industry?: string | null;
  featured?: boolean;
  displayOrder?: number;
}

export interface UpdatePartnerInput {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  industry?: string | null;
  featured?: boolean;
  displayOrder?: number;
}

export interface Service {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  icon: string | null;
  active: boolean;
  /** @nullable */
  activeFrom: string | null;
  /** @nullable */
  activeUntil: string | null;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateServiceInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  active?: boolean;
  /** @nullable */
  activeFrom?: string | null;
  /** @nullable */
  activeUntil?: string | null;
  displayOrder?: number;
}

export interface UpdateServiceInput {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  active?: boolean;
  /** @nullable */
  activeFrom?: string | null;
  /** @nullable */
  activeUntil?: string | null;
  displayOrder?: number;
}

export interface DashboardSummary {
  partnerCount: number;
  featuredPartnerCount: number;
  serviceCount: number;
  activeServiceCount: number;
  recentPartners: Partner[];
  recentServices: Service[];
}