# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Easy Way Agency Admin API
servers:
  - url: /api
    description: Base API path
tags:
  - name: auth
    description: Authenticated user operations
  - name: health
    description: Health operations
  - name: partners
    description: Partner companies management
  - name: services
    description: Services management
  - name: dashboard
    description: Dashboard summaries

paths:
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current authenticated admin status
      description: Returns the current Clerk user identity and whether their email is allowlisted in ADMIN_EMAILS.
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMe"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard summary
      description: Returns counts and recent activity for dashboard overview
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List partner companies
      description: Returns all partner companies, ordered by displayOrder.
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"

  /admin/partners:
    get:
      operationId: listAdminPartners
      tags: [partners]
      summary: List partner companies for admin
      description: Returns all partner companies for the admin dashboard.
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a new partner company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartnerInput"
      responses:
        "201":
          description: Partner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get a single partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/partners/{id}:
    get:
      operationId: getAdminPartner
      tags: [partners]
      summary: Get a single partner for admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePartnerInput"
      responses:
        "200":
          description: Updated partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete a partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      description: Returns all services, ordered by displayOrder.
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /admin/services:
    get:
      operationId: listAdminServices
      tags: [services]
      summary: List services for admin
      description: Returns all services for the admin dashboard.
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/services/{id}:
    get:
      operationId: getAdminService
      tags: [services]
      summary: Get a single service for admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AdminMe:
      type: object
      properties:
        userId:
          type: string
        email:
          type: ["string", "null"]
          format: email
        isAdmin:
          type: boolean
      required:
        - userId
        - email
        - isAdmin

    Partner:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        featured:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - logoUrl
        - websiteUrl
        - description
        - industry
        - featured
        - displayOrder
        - createdAt
        - updatedAt

    CreatePartnerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        featured:
          type: boolean
        displayOrder:
          type: integer
      required:
        - name

    UpdatePartnerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        logoUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        featured:
          type: boolean
        displayOrder:
          type: integer

    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        active:
          type: boolean
        activeFrom:
          type: ["string", "null"]
          format: date-time
        activeUntil:
          type: ["string", "null"]
          format: date-time
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - icon
        - active
        - activeFrom
        - activeUntil
        - displayOrder
        - createdAt
        - updatedAt

    CreateServiceInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        active:
          type: boolean
        activeFrom:
          type: ["string", "null"]
          format: date-time
        activeUntil:
          type: ["string", "null"]
          format: date-time
        displayOrder:
          type: integer
      required:
        - title

    UpdateServiceInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        active:
          type: boolean
        activeFrom:
          type: ["string", "null"]
          format: date-time
        activeUntil:
          type: ["string", "null"]
          format: date-time
        displayOrder:
          type: integer

    DashboardSummary:
      type: object
      properties:
        partnerCount:
          type: integer
        featuredPartnerCount:
          type: integer
        serviceCount:
          type: integer
        activeServiceCount:
          type: integer
        recentPartners:
          type: array
          items:
            $ref: "#/components/schemas/Partner"
        recentServices:
          type: array
          items:
            $ref: "#/components/schemas/Service"
      required:
        - partnerCount
        - featuredPartnerCount
        - serviceCount
        - activeServiceCount
        - recentPartners
        - recentServices