# lib/api-zod/src/generated/types/adminMe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Easy Way Agency Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminMe {
  userId: string;
  /** @nullable */
  email: string | null;
  isAdmin: boolean;
}